+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/EndpointsExample.java
@@ -40,18 +40,18 @@ public class EndpointsExample {
     try (KubernetesClient client = new KubernetesClientBuilder().build()) {
       Namespace ns = new NamespaceBuilder().withNewMetadata().withName(NAMESPACE).addToLabels("this", "rocks").endMetadata()
           .build();
-      logger.info("Created namespace: {}", client.namespaces().createOrReplace(ns));
+      logger.info("Created namespace: {}", client.namespaces().resource(ns).createOrReplace());
       try {
         logger.info("Namespace: {}", ns);
         Deployment deployment = client.apps().deployments().inNamespace(NAMESPACE)
-            .load(EndpointsExample.class.getResourceAsStream("/endpoints-deployment.yml")).get();
+            .load(EndpointsExample.class.getResourceAsStream("/endpoints-deployment.yml")).item();
         logger.info("Deployment created");
-        client.apps().deployments().inNamespace(NAMESPACE).create(deployment);
+        client.apps().deployments().inNamespace(NAMESPACE).resource(deployment).create();
 
         Service service = client.services().inNamespace(NAMESPACE)
-            .load(EndpointsExample.class.getResourceAsStream("/endpoints-service.yml")).get();
+            .load(EndpointsExample.class.getResourceAsStream("/endpoints-service.yml")).item();
         logger.info("Service created");
-        client.services().inNamespace(NAMESPACE).create(service);
+        client.services().inNamespace(NAMESPACE).resource(service).create();
 
         Endpoints endpoints = new EndpointsBuilder()
             .withNewMetadata().withName("external-web").withNamespace(NAMESPACE).endMetadata()
@@ -60,7 +60,7 @@ public class EndpointsExample {
             .endSubset()
             .build();
         logger.info("Endpoint created");
-        client.endpoints().inNamespace(NAMESPACE).create(endpoints);
+        client.endpoints().inNamespace(NAMESPACE).resource(endpoints).create();
         logger.info("Endpoint url");
         endpoints = client.endpoints().inNamespace(NAMESPACE).withName("external-web").get();
         logger.info("Endpoint Port {}", endpoints.getSubsets().get(0).getPorts().get(0).getName());