+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/LoadMultipleDocumentsFromFileExample.java
@@ -17,6 +17,7 @@ package io.fabric8.kubernetes.examples;
 
 import io.fabric8.kubernetes.api.builder.Visitor;
 import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.fabric8.kubernetes.api.model.KubernetesListBuilder;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
 import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
@@ -39,7 +40,7 @@ public class LoadMultipleDocumentsFromFileExample {
     }
     try (KubernetesClient client = new KubernetesClientBuilder().withConfig(configBuilder.build()).build()) {
       List<HasMetadata> list = client
-          .load(LoadMultipleDocumentsFromFileExample.class.getResourceAsStream("/multiple-document-template.yml")).get();
+          .load(LoadMultipleDocumentsFromFileExample.class.getResourceAsStream("/multiple-document-template.yml")).items();
       logger.info("Found in file: {} items.", list.size());
       for (HasMetadata meta : list) {
         logger.info(display(meta));
@@ -47,12 +48,15 @@ public class LoadMultipleDocumentsFromFileExample {
 
       //noinspection Convert2Lambda
       list = client.load(LoadMultipleDocumentsFromFileExample.class.getResourceAsStream("/multiple-document-template.yml"))
-          .accept(new Visitor<ObjectMetaBuilder>() {
-            @Override
-            public void visit(ObjectMetaBuilder item) {
-              item.addToLabels("visitorkey", "visitorvalue");
-            }
-          }).get();
+          .items();
+      KubernetesListBuilder kubernetesListBuilder = new KubernetesListBuilder();
+      kubernetesListBuilder.addAllToItems(list);
+      kubernetesListBuilder.accept(new Visitor<ObjectMetaBuilder>() {
+        @Override
+        public void visit(ObjectMetaBuilder item) {
+          item.addToLabels("visitorkey", "visitorvalue");
+        }
+      });
 
       logger.info("Visited: {} items.", list.size());
       for (HasMetadata meta : list) {